Cherry-pick focus/blur fixes from master (#741)
* Add nullability checks (#704)

* Update RCTCxxBridge.mm

* add nullability checks

* Fix ability to call focus()/blur() on any view (#738)

This fixes a regression from when we upgraded to 0.63 where we could no longer call `focus()` or `blur()` on any `View` (including `Text`). This regression was due to a change that dispatches these as commands (to be compatible with Fabric) instead of directly using `UIManager` to change focus. Only the `TextInput` supported these commands.

* pod install

* Fix ability to call focus()/blur() on any view (#738)

This fixes a regression from when we upgraded to 0.63 where we could no longer call `focus()` or `blur()` on any `View` (including `Text`). This regression was due to a change that dispatches these as commands (to be compatible with Fabric) instead of directly using `UIManager` to change focus. Only the `TextInput` supported these commands.

Co-authored-by: Scott Kyle <scott@appden.com>