[upstream][paper] Tab moves focus from multiline TextInput
Summary: This matches the behavior we already have on Windows. We almost never want the Tab key to actually insert a Tab character, so it makes sense allow the Tab to navigate focus as expected. Even if we did want to use `TextInput` for something like a text editor, the `Tab` key would need to be overridden from JS to insert spaces instead or indent bullets, etc.

Test Plan:
Tab and Shift-Tab work as expected with changes in D28549210

{F617169619}

Reviewers: lyahdav, ackchiu, olgarithm, mowang

Reviewed By: lyahdav

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28548943

Tasks: T87254999

Signature: 28548943:1621473300:71097993927b3add7387f3d5264107085fb055f9