[upstream][paper] Relayout Root View on viewDidEndLiveResize
Summary:
# Problem
Sometimes, at the end of a live resize, the window size is not propagated to resize the content. This leads to situations like this one, where the content is accidentally bigger than the window (see the "Type a message" textbox):
https://www.internalfb.com/intern/px/p/24pVJ/

# Solution
This diff schedules a relayout at the end of a live resize to ensure that the latest value is used.

# How did I discover this?
First I added [logging to product code here](https://www.internalfb.com/code/fbsource/[fbdb38f5d118608951d4807629af599b638f6b74]/xplat/archon/packages/rn/src/messenger/ui/chatNavigation/ChatNavigationContainer.tsx?lines=100) and determined that the last window dimensions value logged is always correct.

I then added a lot of native logging to log the sizes of the RootView and RootContentView, and the events being fired. (Full diff with all my logging in V0.1 of D36719158)

From the logging I determined that the problem ALWAYS occurred when `onWindowDidEndLiveResize` was the most recent event fired. The size in that event was always correct, but [`setAvailableSize`](https://www.internalfb.com/code/fbsource/[b530355767b24e87c64e8197fae2475133bede2f]/xplat/js/react-native-github/React/Modules/RCTUIManager.m?lines=393) was not called. `onWindowDidResize` events did result in `setAvailableSize` being called.

From googling this problem I discovered the overridable method [`viewDidEndLiveResize`](https://developer.apple.com/documentation/appkit/nsview/1483543-viewdidendliveresize).

With this fix, the size always ends up correct at the end and `setAvailableSize` is called after `onWindowDidEndLiveResize` if the size changed.
# Other things I tried
Pulling in https://github.com/microsoft/react-native-macos/pull/439 does seem to fix the issue (as in it can't be reproduced anymore), but makes resizing look extremely janky.

# Other things I tried that did not work
I was still able to reproduce the issue doing these. Adding these here for posterity.
* Removing all event coalescing
* Removing event coalescing for just onWindowDidEndLiveResize
* Removing event coalescing for just onWindowDidResize
* Removing the synchronous relayout [here](https://www.internalfb.com/code/fbsource/[b530355767b24e87c64e8197fae2475133bede2f]/third-party/microsoft-fork-of-react-native/react-native/React/Modules/RCTUIManager.m?lines=421)
* Sending an onWindowDidResize event at the same time as the onWindowDidEndLiveResize event [here](https://www.internalfb.com/code/fbsource/[b530355767b24e87c64e8197fae2475133bede2f]/xplat/archon/native/rn/src/NativeWindowReact-macOS.mm?lines=587)
* Set `autoresizingmask` in RCTRootContentView, and set `self.autoresizesSubviews = YES` and `self.clipsToBounds = YES` in RCTRootView.

Full diff with all my logging and fix attempts in V0.1 of D36719158

Test Plan:
Logged the size of the Root View on every resize and end of live resize to ensure they were the correct value. See logging in V0.1 of D36719158

Tried to trigger the bug for a long time (minutes) when before this diff it can trigger pretty quickly.

A short video of the window getting "stuck" in the wrong size, and then fixing itself at the end of the live resize:
https://pxl.cl/257PV

NOTE: There are several resizes that happen on the way to the corrected layout. This is because of the way layout is calculated in `ChatNavigationContainer.tsx`. They are using [debounce to wait 1s to resize the left panel](ChatNavigationContainer.tsx). I have some thoughts on how to improve `ChatNavigationContainer.tsx` in the comments of T107249807.

A longer video showing many resizes:
https://pxl.cl/257Q9

Reviewers: skyle, lyahdav

Reviewed By: skyle

Subscribers: joesavona, generatedunixname499725568, kacieb, laurentan

Differential Revision: https://phabricator.intern.facebook.com/D36785342

Tasks: T107249807

# Conflicts:
#	React/Base/RCTRootView.m