[upstream][[paper] remove drag target from Image
Summary: The purpose of this diff is to unregister the drag on images on Mac os. The main reason is if we add a drag target on component which contains the image like Grid call view on react side, then drag functionality won't work because it is interfering with drag and drop on react side with the drag and drop on native mac os in RCT Image

Test Plan:
Before :

As you could see the drag and drop is not working initially when the image background is present.

https://pxl.cl/2bnnF

After :

https://pxl.cl/2bnpj

Reviewers: lyahdav, shawndempsey, zaza, mowang, #rp_content

Reviewed By: zaza

Subscribers: ilit

Differential Revision: https://phabricator.intern.facebook.com/D38863731

# Conflicts:
#	Libraries/Image/RCTImageView.mm