[upstream][paper] Catch and broadcast outside window mouse up events as opt-in
Summary:
Native controls use event tracking loops which keep the touch handler from receiving the mouse up event that stops the event tracking.

By swapping the NSApplication event funnel method with a custom override, we can catch the events without having to extend the NSApplication class.

The override catches mouse up events outside app windows and broadcasts a notification with the event to all touch handlers so that they have a chance of updating their state correctly, submitting the completed mouse down-up cycle.

This feature was implemented as an opt-in to allow testing the effect of this fix through a QE to check for regressions.

This would allow us to remove all the manual contextual menu notification that was used previously for the missing right mouse up event.

Test Plan:
Initialize `_notifyOutsideViewEvents` to `YES`.

Run rn-tester. Click inside a textfield and drag the mouse out of the window, releasing the mouse button once outside.

Click on a control. With the improved event tracking, the control should receive the first press event.

Fix disabled (click on "Button" does not work):
https://pxl.cl/2tQPf

Fix enabled (click on "Button" works):
https://pxl.cl/2tQP7

Reviewers: shawndempsey, chpurrer

Reviewed By: chpurrer

Subscribers: generatedunixname89002005327315

Differential Revision: https://phabricator.intern.facebook.com/D43558607

Tasks: T119232032