[paper][touchhandler] Fix mouse dragging events being flagged as duplicate events
Summary:
Update the duplicate event detection logic to allow mouse drag events to be sent out continuously.

Mouse drag events maintain the same event type and event number, but update the location of the event. This diff checks for location updates so that mouse drag events wouldn't be flagged as duplicates.

Test Plan:
Launch RNTester, click and drag in and out of a Pressable in the Pressable demo.

Without the fix:
https://pxl.cl/2wq4R

With the fix:
https://pxl.cl/2wq4W

Reviewers: shawndempsey, chpurrer

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D43985809

Tasks: T147732113