[upsream][paper] Horizontal ScrollView should only have scrollbar padding when "Always" showing
Summary:
**Context**

https://fb.workplace.com/groups/messengerdesktopfeedback/posts/1538647899953946/

Extra padding in horizontal ScrollViews were visible when the macOS scrollbar setting was set to "Always".

We calculate the padding for the scroll indicators for "NSScrollerStyleLegacy", this is when the scroll indicator is always visible.

We need to reset this padding to zero when the scroll indicator style changes to "Overlay"

**Change**

- reset the scroll indicator padding to zero when the scroll indicator style is not "Legacy"

Test Plan:
|RNTester - horizontal scrollview should only have padding when Always showing scrollbar|
|https://pxl.cl/2wMbM|j

|before|
| https://pxl.cl/2wMr2 |

|after|
|https://pxl.cl/2wMpV|

Reviewers: chpurrer, lefever, helenistic

Reviewed By: helenistic

Differential Revision: https://phabricator.intern.facebook.com/D44063967

Tasks: T147915593

# Conflicts:
#	React/Views/ScrollView/RCTScrollContentView.m