[upstream][paper] Fix onScroll not being emitted when scrolling contents of RCTScrollView.
Summary:
The last fix filtering out duplicate onScroll events was using the documentView.frame to read the scroll position, while the actual source of truth for the scroll position is the NSScrollView's NSClipView bounds.

This diff updates the fix to read and store the last scroll position by getting the bounds from `_scrollView.contentView` (the NSClipView).

Test Plan:
Run RNTester and use the ScrollView example to check that onScroll events are logged from the JS side.

Without the fix (no onScroll events logged):
https://pxl.cl/2Lrbv

With the fix (onScroll events are logged):
https://pxl.cl/2LrbM

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D46655947

Tasks: T154171955, T155495647

# Conflicts:
#	packages/react-native/React/Views/ScrollView/RCTScrollView.m