[upstream][paper] Fix RCTScrollView duplicate onScroll event filter
Summary:
`RCTScrollView` was submitting `onScroll` events after an `onContentSizeChanged` event. This was caused by the `setFrame` call leading to a `DocumentViewBoundsChanged` notification that would get triggered a second time after the horizontal/vertical scrollers would be enabled/disabled.

This double notification would trigger a `scrollViewDidScroll` call with an `onScroll` event being sent because the scroll position would toggle temporarily due to the scrollers being shown/hidden.

This diff checks the scroll position with the last sent position after the scrollers were set. This fixes the wrong `onScroll` event during the `setFrame` because the scroll position is stable and valid by the time we do the comparison.

Test Plan:
Run Workplace Chat and toggle between threads quickly.

Without the fix an onScroll event with a contentSize of 0 gets submitted after the onContentSizeChanged event. This leads to the VirtualizedList not rendering the last submitted messages and the "fill" messages not being loaded because the MessageScrollContainer never gets notified of the render of the last provided messages.

With the fix, the messages are loaded correctly on every thread switch:
https://pxl.cl/2NLpl

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D46817106

Tasks: T155495647