[upstream][paper] Add back unconditional RCTSwitch onChange call
Summary:
This diff removes the conditional `onChange` call for the `RCTSwitch`. On macOS we register an action on the switch, which triggers every time the user clicks on the component. On iOS it is on value changes that the registered method is called, which required filtering out programmatic value settings for the `onChange` call.

This fixes a bug on macOS where the first toggle was not emitting a value change on the JS side because the native value was not being provided and at https://fburl.com/code/mztyhsn8 `shouldUpdateNativeSwitch` would be false due to `native.value` not having the latest assigned value.

The latest value is not available because setting the value programmatically on the native switch does not trigger an `onChange` on the native side, which does not update the native state on the JS side of the component.

Test Plan:
In messenger desktop, open the settings (cmd+,):
- select *Active status*
- toggle the *Show active status* switch to off

Without the fix, the setting value stays on:
 https://pxl.cl/2V1p2 #rn

With the fix, the setting value gets updated to off:
 https://pxl.cl/2V1mr

Reviewers: shawndempsey, chpurrer, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D47235602

Tasks: T156924979