[upstream][paper] Fix missing RCTTextView element in accessibility hierarchy
Summary:
The accessibility hierarchy is automatically derived from the view hierarchy. Yet in the case of RCTTextView, the child RCTUnfocusableTextView was skipping the parent view when building the hierarchy.

The best guess for this is the use of `canDrawSubviewsIntoLayer` which may break the accessibility hierarchy tree building in the latest version of AppKit.

This diff sets the parent manually after adding RCTUnfocusableTextView as a child to RCTTextView so that the accessibility hierarchy would include the RCTTextView which is holding all the react accessibility properties (e.g. `testID`) which are used by tests and expected to be present.

Test Plan:
Run messenger desktop and use the accessibility inspector to query the accessibility hierarchy of the `Chats` text in the left side bar.

With the change, the RCTTextView element is present in the accessibility hierarchy.

 {F1045363538}

Reviewers: shawndempsey, ericroz, #rn-desktop, #zeratul

Reviewed By: ericroz

Differential Revision: https://phabricator.intern.facebook.com/D47336932

Tasks: T157739043