[upstream][paper] Only focus textfields when it effectively moved to a window
Summary:
Due to a bad merge in 0.68, the `RCTView` move to window call got moved to `updateClipViewBoundsObserverIfNeeded` which gets called by mouse enter and leave notification methods.

This change puts the method back into the `viewDidMoveToWindow` method and also updates the text input handling of the autofocus to check if the text input effectively moved to a window before marking it as being shown.

These changes fix the autofocus for text inputs.

Test Plan:
Select a message thread in Messenger Desktop and check that the composer is focused automatically.

Without the fix, the composer is not focused when selecting a thread:
https://pxl.cl/32TQt

With the fix, the composer is focused:
https://pxl.cl/32TR1

Reviewers: chpurrer, #rn-desktop, #zeratul, #messenger_desktop_experience

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D47795829