[upstream][paper] Set up the correct accessibility hierarchy for Text components
Summary:
With the use in RN 0.71 of the new `RCTUnfocusableTextView` subview for rendering the text contents in the `RCTTextView`, the a11y hierarchy was rendering the `RCTUnfocusableTextView` as a non-hittable static text.

This breaks the e2e tests that click on buttons through a selector targeting the child text contents of the button (e.g. `Button :text("My button label")`). The text being non-hittable, no clicks would be supported on the text itself to simulate a click on the parent button.

This diff hides the `RCTUnfocusableTextView` from the a11y hierarchy of the Text component and marks the parent `RCTTextView` as an accessibilityElement so that it would be accessible and "hittable" in e2e tests and also include any testIDs which are assigned to the `RCTTextView` a11y element.

This restores the original a11y hierarchy of the Text component and hides the `RCTUnfocusableTextView` which is internally used only for rendering anyway and therefor shouldn't be exposed in the a11y hierarchy.

Test Plan:
Start Messenger Desktop.
From Xcode, go to Xcode > Open Developer Tools > Accessibility Inspector.

Target a thread in the threads list. The a11y inspector hierarchy shows the button text as a single (text) [RCTTextView] element.

 {F1062756775}

Reviewers: shawndempsey, chpurrer, bumbu, #rn-desktop, #zeratul, #messenger_desktop_experience

Reviewed By: bumbu

Differential Revision: https://phabricator.intern.facebook.com/D47979435

Tasks: T157739043