[upstream][paper] Implement double click callback on RCTView
Summary:
The Zeratul double click on the title bar didn't trigger the zoom action set up in the product code. This was not working due to the removal of the double click event handling on RCTView.

This diff adds back support on macOS for double clicks within an RCTView.

Test Plan: Tested later in this stack.

Reviewers: shawndempsey, chpurrer, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D48670208

Tasks: T158571385

# Conflicts:
#	packages/react-native/React/Views/RCTView.h