[upstream][paper] Fix laggy scrolling on macOS Sonoma 14.0
Summary:
AppKit updated the default clipsToBounds value for NSView to NO starting from macOS Sonoma 14.0. This resulted in the NSView visibleRect method to return a region that could be larger than the view bounds.

This would cause the mouseEnter/Leave hit testing to run on all views requiring mouse region tracking, which would make scroll views laggy in Workplace Chat.

This checks the mouse position against the view bounds on top of the visible rect so that we'd check if we enter/leave the bounds of the view itself.

Test Plan:
Scroll through the message list in Workplace Chat on macOS Sonoma 14.0.

Check that the scrolling is smooth.

https://pxl.cl/3wbFz

Reviewers: shawndempsey, ericroz, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D49941424

Tasks: T164996779