[fabric][touchhandler] Add event broadcasting to surface touch handler
Summary:
This diff adds a helper method that can broadcast and event to all surface touch handlers.

Events happening outside the view hierarchy of the surface touch handler have to be broadcast to all surface touch handler so that each instance can update the internal state based on the mouse up event.

Test Plan: Tested later in this stack (D50480942)

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D50480946

# Conflicts:
#	React/Fabric/RCTSurfaceTouchHandler.mm