[fabric][touchhandler] Remove manual context menu notification to surface touch handler
Summary:
With the changes applied in this stack there is no need for notifying the surface touch handler of contextual menu's being displayed.

The event catching and broadcasting is able to catch the breaking event swallowing and cancel the ongoing gesture started by the right mouse down.

Test Plan:
Force enable Fabric in Zeratul by setting the `useFabric` property to `YES` on the created `ReactRootViewManager` instances.

- Right click in the main window.
- Toggle between appearance styles in the settings more than 11 times.
- Right click on a thread and click away.

 https://pxl.cl/3DhQ1

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D50480942