[fabric] Show autohiding scrollbars by default for RCTScrollViewComponentView
Summary:
This diff enables the scrollbars on init for scroll views in Fabric together with autohiding.

With the support of show/hide scroll bars properties, this adds full support for scroll bar visibility settings to scroll views in Fabric.

Test Plan:
Run Zeratul with Fabric enabled and scroll a message thread.
With the change the scrollbars are showing up as expected.

| Before | After |
|--|
|  https://pxl.cl/3MwMl  |   https://pxl.cl/3MwNJ  |

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D51165365