[fabric] Support contentInset on RCTScrollViewComponentView
Summary:
This diff implements the `contentInset` prop setting for the native scroll view used by `RCTScrollViewComponentView` in Fabric.

The diff clips the scrollers and set the content inset on the native NSScrollView to match the functionality in UIKit.

Test Plan:
Run Zeratul with Fabric enabled. Scroll to the start of a message thread. With the change, the scroll content is correctly clipped to the top of the visible region of the message thread.

| Before | After |
|--|
|  https://pxl.cl/3MwQ2 |  https://pxl.cl/3MwQf |

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D51165364