[fabric] Implement setContentOffset on Fabric scroll view to support scrollTo and scrollToEnd commands
Summary:
This diff adds the setContentOffset method to the `RCTEnhancedScrollView` used by `RCTScrollViewComponentView` in Fabric.

This method is used by the `scrollTo` and `scrollToEnd` commands to programmatically scroll the view from the JS side.

Test Plan:
Run Zeratul with Fabric enabled and click on the arrow in the message thread to scroll to the end of the thread.

| Before | After |
|--|
|  https://pxl.cl/3MVQn  |  https://pxl.cl/3MVQs  |

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D51222018

Tasks: T157893014