[fabric] Implement zoomToRect on Fabric scroll view
Summary:
This diff adds the zoomToRect method to the `RCTEnhancedScrollView` used by `RCTScrollViewComponentView` in Fabric.

This method is used by the `zoomToRect` command provided to the ScrollView on the JS side.

Test Plan:
Run Zeratul with Fabric enabled, open an image and use the zoom function in the media viewer.

| Before | After |
|--|
|   https://pxl.cl/3MVSl   |   https://pxl.cl/3MVSd   |

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D51222020

Tasks: T169758894