[fabric] Implement flashScrollIndicators on Fabric scroll view
Summary:
This diff adds the flashScrollIndicators method to the `RCTEnhancedScrollView` used by `RCTScrollViewComponentView` in Fabric.

This method is used by the command with the same name to enable the flashing of the scroll bars from the JS side.

Test Plan:
Run Zeratul with Fabric enabled and with a custom VirtualizedList.js implementation that adds the following code to `_onContentSizeChange`:
```
this.getScrollResponder().flashScrollIndicators();
```

This should flash the scroll bars every time the content size of the scroll view changes. Meaning adding a message to a thread should flash the scroll bars using the `flashScrollIndicators` command.

 https://pxl.cl/3MVZ1

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D51222019