[fabric] Implement escape/cancel key press callback for TextInput
Summary:
Pressing the escape key down in a TextInput on macOS will be captured on the native side as a cancel event.

This diff adds support for sending the "Escape" key as a key pressed event and wires the delegate method of the backing text input to call the key press event handler with the correct payload.

Test Plan:
- Run Zeratul with Fabric enabled
- Add a key press handler on the textfield logging the received data (e.g. on MDSTextInput)
- Focus the textfield
- Press the escape key
- The key press handler should be called and print that the "Escape" key was pressed.

 https://pxl.cl/3XHR2

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D52005857

Tasks: T167538822