[fabric] Fix random TextInput cursor position changes while typing
Summary:
When text completions or special character symbols are displayed ahead of time, the content is present in the native text view attributed string.

If a state update is applied, the cursor position will not be read as being at the end of the string while typing, which will shift the cursor position after the update.

This diff disables native text view content updates while temporary context is displayed to fix the wrong cursor position updates.

Test Plan:
Run Zeratul with Fabric and type text in the message composer until macOS renders a text completion/suggestion. Keep typing and see if the cursor stays at the end of the string.

| Before | After |
|--|
|  https://pxl.cl/49PW4  |  https://pxl.cl/49PWh  |

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D52787094

Tasks: T174286406