[fabric] Add mouse enter/leave tracking to the View component
Summary:
This diff implements mouse enter/leave tracking for the area covered by the View component.

The tracking is handled by configuring a tracking area on the NSView when either of the handlers is set on the view. This enables the NSResponder `mouseEntered:` and `mouseExited:` notifications which are translated to mouse events.

Test Plan:
* Run Zeratul with Fabric enabled.
* Move the cursor over controls that have hover-dependent styling.

 https://pxl.cl/4jSw5

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D53529015

Tasks: T154617556