[fabric] Adding a getter to layoutable shadow node to mark vertical axis flipping use
Summary:
To correctly transform shadow node layout metrics to ancestor coordinate spaces on macOS, we need to take view flipping into account to correctly flip the vertical axis during the transform.

This diff adds a virtual function to `LayoutableShadowNode` that can be overridden to indicate if the view linked to the shadow node will layout its children with a flipped vertical axis (y = 0 is a the bottom of the view, the y axis is pointed upwards).

By default the method returns false, which is the common case for native views in RN. On macOS, all views extending `RCTUIView` will configure the vertical axis to match the default on iOS. The origin is located at the top of the view and the y axis is pointed downwards.

Test Plan: Tested later in this stack.

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D54931344

Tasks: T182039265, T182037720

Tags: uikit-diff