[fabric] Support flipped views in relative layout metrics evaluation
Summary:
This diff updates the implementation of `LayoutableShadowNode::computeRelativeLayoutMetrics` which is used by the React `measure` api to return the layout metrics of a shadow node relative to an ancestor node.

If a shadow node is a child of shadow node that lays out its children with a flipped vertical axis, then all transform calculations based on that node have to happen with coordinates within that same axis.

The diff compares the flipped view state between the current node and the parent node to convert the current origin being transformed to the right axis orientation.

A conversion happens if:
- we move from a standard to a flipped axis
- we move from a flipped axis to a standard one

Test Plan: Tested later in this stack.

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Differential Revision: https://phabricator.intern.facebook.com/D54931339

Tasks: T182039265, T182037720