[fabric] Add a custom shadow node to ScrollContentComponentView
Summary:
This diff adds a custom shadow node for the ScrollContentView component, allowing to propagate the view flipping state to the shadow tree.

The ScrollContentView was using a generated shadow node. The generation of the descriptor and shadow node was disabled in the spec and a custom implementation was added for the component.

Test Plan: Tested later in this stack.

Reviewers: shawndempsey, #rn-desktop

Reviewed By: shawndempsey

Subscribers: ramanpreet

Differential Revision: https://phabricator.intern.facebook.com/D54931340

Tasks: T182039265, T182037720